+++ examples/redis-async.rs
@@ -38,7 +38,7 @@ async fn cached_sleep_secs(secs: u64) -> Result<(), ExampleError> {
     map_error = r##"|e| ExampleError::RedisError(format!("{:?}", e))"##,
     ty = "cached::AsyncRedisCache<u64, String>",
     create = r##" {
-        AsyncRedisCache::new("cache_redis_example_cached_sleep_secs", 1)
+        AsyncRedisCache::new("cache_redis_example_cached_sleep_secs", Duration::from_secs(1))
             .set_refresh(true)
             .build()
             .await
@@ -67,7 +67,7 @@ static CONFIG: Lazy<Config> = Lazy::new(Config::load);
     map_error = r##"|e| ExampleError::RedisError(format!("{:?}", e))"##,
     ty = "cached::AsyncRedisCache<u64, String>",
     create = r##" {
-        AsyncRedisCache::new("cache_redis_example_cached_sleep_secs_config", 1)
+        AsyncRedisCache::new("cache_redis_example_cached_sleep_secs_config", Duration::from_secs(1))
             .set_refresh(true)
             .set_connection_string(&CONFIG.conn_str)
             .build()