+++ src/proc_macro.rs
@@ -226,6 +226,7 @@ fn keyed(a: &str, b: &str) -> usize {
 
 ```rust,no_run
 use cached::proc_macro::once;
+use std::time::Duration;
 
 /// Only cache the initial function call.
 /// Function will be re-executed after the cache